Fix the mess by removing the explicit base font size declaration 😌
As bootstrap assumes the base font-size to be default 16px so, setting
the base font-size to 50% was a bad idea. To clean everything up, this
commit change everything according to base browser font-size (16px).
An additional mixin was added and used to convert pixels to rems. Some
small tweaks were also made.